Accept Timestamp<Unix> in DateTime's public API
Summary:
This adds a small subset of functionality from D7008450 so that the public API of DateTime can be typed as accepting/returning `Timestamp<Unix>`. This will ensure that users are always passing a timestamp of the correct unit and of the right source.

Hopefully, since these conversion functions are literally as simple as possible, they'll be inlined.

Reviewed By: fredemmott

Differential Revision: D7826617

fbshipit-source-id: 2be77838425448ff9da8c12a1b39ab9e8dfce098