Rebuild Thrift and fix hack errors
Summary:
This is a change from Indexish to KeyedContainer and requires a really simple codemod to fix definitions of the handler classes. This is all fine because these classes are now effectively identical.

This also requires a simple addition to lies_we_tell_hack but nothing for hhvm.

allow-large-files
allow_many_files
ignore-nocommit
bypass-lint

Reviewed By: yfeldblum, rodmk

Differential Revision: D13938016

fbshipit-source-id: 008b442b8391b04a6349fecd4660580d928f8359