Run DataProviderTestCodemod to improve typings on DataProviders
Summary:
See previous diff (D16707810) for explanation on how this works.

`phps DataProviderTestCodemod -v`

~~There are a LOT of lint warnings.  Some of these are hackfmt.  For now, I have the formatting changes in D16735456 (4/x), because it blows up the lines of code and makes it harder to spot check.  Will run `hg diff -c -1 | hackfmt --diff` and merge into this diff before landing.~~ Have merged hackfmt changes into this diff to land.

Some of the lint warnings are related to invalid values created by the codemod, saying that the type of the provider does not match the type of the function.  These are correct.  These are cases where the values returned by the provider actually don't match the type of the function.  Leaving these warnings to be cleaned up by others in individual cases when they can take them a file at a time.

allow-many-files
drop-conflicts

Reviewed By: kmeht

Differential Revision: D16733392

fbshipit-source-id: ca1526afcb30485d20de4d134eef6c261631f0b4