Remove "Experimental\\" from HSL IO, OS namespaces. No change to status.
These libraries are still experimental - this remains indicated by the
package name, or by the internal path and lack of aliasing in flib.

Renaming as:
- the namespace indication is not necessary
- it makes moving things out of experimental a breaking change
- it's very awkward for namespaces that contain builtins and
non-builtings, e.g. `OS\`

Only doing this for IO/OS stuff for now as these are the only ones that
have builtins, and are under active development.