Remove "Experimental\\" from HSL IO, OS namespaces. No change to status. (#116)
Summary:
These libraries are still experimental - this remains indicated by the
package name, or by the internal path and lack of aliasing in flib.

Renaming as:
- the namespace indication is not necessary
- it makes moving things out of experimental a breaking change
- it's very awkward for namespaces that contain builtins and
non-builtings, e.g. `OS\`

Only doing this for IO/OS stuff for now as these are the only ones that
have builtins, and are under active development.
Pull Request resolved: https://github.com/hhvm/hsl-experimental/pull/116

Reviewed By: jjergus

Differential Revision: D20678984

Pulled By: fredemmott

fbshipit-source-id: bb067594f2b77ed75735dc6dbf2a9e528ac9aeeb