Remove disposable interfaces, disposable-first approach.
Discussed when removing buffered operations like `readLine`; we're going
to be replacing those with a `BufferedReader($handle)` class or similar,
so composition is going to be a key part of usage. Composition with
disposables is banned, so make the API match the common case.

This also cuts the number of interfaces roughly in half.

Future work (very soon): add a disposable API for autoclose. I'm leaving
that for a later PR as that's more bikesheddable.

Some options include:
- a proxying wrapper that implements the Handle APIs
- something like the new TemporaryFile API (with `getHandle`)
- `$f = File\open_...(); using $f->scopeGuard();`
Whichever we go with, I'll adjust TemporaryFile to match, but want to
leave that for later diff.