Remove disposable interfaces, disposable-first approach. (#130)
Summary:
Discussed when removing buffered operations like `readLine`; we're going
to be replacing those with a `BufferedReader($handle)` class or similar,
so composition is going to be a key part of usage. Composition with
disposables is banned, so make the API match the common case.

The short version of that discussion is that combining async with buffered IO is complicated, and a strict layering makes things somewhat clearer and safer.

This also cuts the number of interfaces roughly in half.

Future work (very soon): add a disposable API for autoclose. I'm leaving
that for a later PR as that's more bikesheddable.

Some options include:
- a proxying wrapper that implements the Handle APIs
- something like the new TemporaryFile API (with `getHandle`)
- `$f = File\open_...(); using $f->scopeGuard();`
Whichever we go with, I'll adjust TemporaryFile to match, but want to
leave that for later diff.
Pull Request resolved: https://github.com/hhvm/hsl-experimental/pull/130

Reviewed By: jjergus

Differential Revision: D21539877

Pulled By: fredemmott

fbshipit-source-id: c883ea3a03505eb6499ab9ccb6d8b8a0ebbe438a