Provide new 'hhvm/hsl-io' virtual package
This can follow SEMVER, and allows users to specify what version of
HSL-IO APIs they are compatible with, rather than new BC-incompatible releases
automatically breaking all current users (due to the library as a whole
following HHVM versioning)

WE ARE NOT YET COMMITTING TO SUPPORTING THIS VIRTUAL PACKAGE.

I'm adding it now so we can test if it solves real-world problems. It
may be removed in future releases.