Fix crazy default memory usage by `genReadAll`/`readAllAsync`
Summary:
If unspecified, there's no limit, so we set to INT64_MAX. There's a catch though - ultimately,
`genRead`/`readAsync` is POSIX `read()` with some awaits, and `read` needs a preallocated buffer.

It ends up being 6GB because the runtime reduces it to the maximum size of a string, which is 6GB.

Read in chunks regardless of size.

Related todo : default buffer size should be replaced by checking the block size of the handle.

Reviewed By: jano, alexeyt

Differential Revision: D25379885

fbshipit-source-id: f0c89d7b868e7c6a8c07b4d6f0fab51c4902537e